Updated error conditions
